vfio-ioctls: do not leak hypervisor specific DeviceFd type
Previously the exported DeviceFd type was dependent on whether KVM or
MSHV was enabled.

This presents a challenge for producing a unified Cloud Hypervisor
binary for both KVM and MSHV. Also I've realized leaking dependent types
are annoying to library users in general because they need to have
different code paths for different configurations.

Provide a wrapper type DeviceFd. User can wrap either a KVM DeviceFd or
a MSHV DeviceFd in it. Provide two constructors for the
VfioContainerDeviceHandle type so that its construction is type-safe.

Although I tried to keep things as close as before, this is still a
breaking change. Users are now expected to use the two new constructors
to build a new VfioContainerDeviceHandle.

Signed-off-by: Wei Liu <liuwe@microsoft.com>