server: pass file by value rather than reference in Backend::dma_map
If the backend wants to hang on to the file descriptor, it now has to
resort to unsafe code to dup() the File object.

Since we don't intend to hold on to the file object, we can just pass
it on by value and let the callee do with it as they please.

While we are here, also reject malformed input with too many file
descriptors.

Signed-off-by: Julian Stecklina <julian.stecklina@cyberus-technology.de>